requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
scikit-image==0.21.0
requests==2.31.0
flask==2.3.3
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.67.1
scikit-learn==1.3.0